requirements.txt
caikit==0.3.1
requests==2.30.0
Pillow==9.5.0
timm==0.9.1
# backend only
transformers[torch]==4.27.4
# frontend only
gradio==3.29.0
grpcio==1.53.0
